requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
python-dotenv>=1.0
requests>=2.31
